Refactor the damage pipeline
